requirements.txt
# ============================================================
# Healthcare RAG - Ask My Docs
# requirements.txt
# ============================================================

# --- Core LLM & Orchestration ---
anthropic>=0.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-anthropic>=0.1.0

# --- Vector Store ---
chromadb>=0.5.0

# --- Document Processing ---
pymupdf>=1.24.0          # PDF parsing (fitz)
pdfplumber>=0.11.0       # Fallback PDF parser
tiktoken>=0.7.0          # Token counting (cl100k_base)

# --- Embeddings & Reranking ---
sentence-transformers>=3.0.0

# --- Sparse Retrieval ---
rank-bm25>=0.2.2

# --- Evaluation ---
ragas>=0.1.21
datasets>=2.20.0

# --- Config & Environment ---
pyyaml>=6.0.2
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0

# --- Logging ---
loguru>=0.7.2

# --- Testing ---
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0

# --- Utilities ---
numpy>=1.26.0
tqdm>=4.66.4
